# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Relay Console API
  version: 0.1.0
  summary: Original control plane API for prompts, runs, traces, datasets, and agents.
servers:
  - url: http://localhost:8000/v1
...
paths:
...
  /workspaces/{workspaceId}/api-keys:
    get:
      tags: [API Keys]
      operationId: listApiKeys
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: API keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyListResponse"
    post:
      tags: [API Keys]
      operationId: createApiKey
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateApiKeyResponse"
  /workspaces/{workspaceId}/api-keys/{apiKeyId}/revoke:
    post:
      tags: [API Keys]
      operationId: revokeApiKey
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/ApiKeyId"
      responses:
        "200":
          description: API key revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Relay-Key
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
    ApiKeyId:
      name: apiKeyId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        key_prefix:
          type: string
        revoked_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, label, key_prefix, created_at]
    ApiKeyListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ApiKey"
      required: [items]
    CreateApiKeyRequest:
      type: object
      properties:
        label:
          type: string
        scopes:
          type: array
          items:
            type: string
        expires_at:
          type: string
          format: date-time
          nullable: true
      required: [label]
    CreateApiKeyResponse:
      type: object
      properties:
        api_key:
          $ref: "#/components/schemas/ApiKey"
        secret:
          type: string
      required: [api_key, secret]
...